Factor out asset extraction tests into separate per-filter suites
